server/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
pymupdf
anthropic
python-dotenv
psycopg2-binary
